copy REST service that commited on sprint_2

+++ app/Console/Commands/rajaongkirFetchKota.php
@@ -6,6 +6,8 @@ use Illuminate\Console\Command;
 use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Client;
 use App\Models\City;
+use App\Repositories\BaseRepository;
+use App\Repositories\CityRepository;
 
 class rajaongkirFetchKota extends Command
 {
@@ -14,7 +16,7 @@ class rajaongkirFetchKota extends Command
      *
      * @var string
      */
-    protected $signature = 'rajaongkir:fetch-kota';
+    protected $signature = 'rajaongkir:fetch-kota {--city_id=} {return-data=false}';
 
     /**
      * The console command description.
@@ -31,11 +33,14 @@ class rajaongkirFetchKota extends Command
     protected $client;
     protected $cityModel = \App\Models\City::class;
     protected $storedCity = [];
+    protected $cityRepository;
 
-    public function __construct()
+    public function __construct(CityRepository $cityRepository)
     {
         parent::__construct();
         $this->client = new Client();
+        $this->cityRepository = $cityRepository;
+
     }
 
     /**
@@ -45,24 +50,30 @@ class rajaongkirFetchKota extends Command
      */
     public function handle()
     {   
-        $result = $this->client->get(
-            config('custom.api.rajaongkir.url').'/city',[
-                'query'=>[
-                    'key'=>config('custom.api.rajaongkir.key')
-                ]
-            ]
-        );
-
-        if($result->getStatusCode()!=200){
-            echo "Request Error $result->getStatusCode()";
+        if($this->cityRepository->isDirectRequest()==false){
+            $this->error('ONLY SUPPORT FOR DIRECT REQUEST');
             return false;
         }
 
-        $dataReturn = json_decode($result->getBody()->getContents());
-        $dataReturn = $dataReturn->rajaongkir->results;
+        $return = $this->cityRepository->getByCityId($this->option('city_id'));
 
-        echo $this->truncateCity();
-        echo $this->storeCity($dataReturn);
+        if( $return == false ){
+            $this->error('REQUEST FAILED');
+            return false;
+        }
+
+        if( (int) $this->option('city_id') > 0 ){
+            $this->info("Uncovered Single Store Data");
+            print_r($return);
+            return false;
+        }
+
+        if($this->argument('return-data')=='return-data'){
+            return $return;         
+        }
+
+        $this->confirmTruncate();
+        $this->storeCity( (array) $return);
         return true;
     }
 
@@ -82,7 +93,7 @@ class rajaongkirFetchKota extends Command
 
         }
 
-        return "TOTAL City Fetch / Stored : $totalData / $totalStore".PHP_EOL;
+        $this->info("TOTAL City Fetch / Stored : $totalData / $totalStore".PHP_EOL);
     }
 
     public function stup($input){
@@ -96,12 +107,13 @@ class rajaongkirFetchKota extends Command
         return $city;
     }
 
-    public function truncateCity()
+    public function confirmTruncate()
     {   
-        $city = new $this->cityModel;
-        $city->truncate();
-
-        return "TABLE CITY TRUNCATED".PHP_EOL;
+        if( $this->confirm('Do you wish to truncate CITY table? [yes|no]') ){
+            $city = new $this->cityModel;
+            $city->truncate();
+        }
+        $this->info("TABLE PROVINCE TRUNCATED".PHP_EOL);
     }
 
     public function getStoredCity(){
